requirements.txt
fastapi
uvicorn
face_recognition
numpy
python-multipart
faiss-cpu
python-jose[cryptography]
passlib[bcrypt]
